# Optimize conversion of ModifierKeys enum from/to string, reduce allocations