--- a/src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Input/ModifierKeysConverter.cs
+++ b/src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Input/ModifierKeysConverter.cs
@@ -2,8 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.ComponentModel;
-using System.Globalization;
+using System.Runtime.CompilerServices;  // for MethodImplAttribute
+using System.ComponentModel;            // for TypeConverter
+using System.Globalization;             // for CultureInfo
 
 namespace System.Windows.Input
 {