Fixes #7502

## Description

Parsing optimization for `ModifierKeys` and its conversion from/to string. Conversion **from** is now alloc-free, conversion to is either free or reduced by a big margin (see benchmarks).

- `ConvertFrom` is now fully flying on `ReadOnlySpan<char>` instead of multiple substrings and trims on string instances.
- `ConvertTo` is now within 2 procs, fast-path and slow-path. The only allocated string instance is with multiple modifiers.
- Strings in `ConvertTo` will now be interned with the rest since we use `StringComparison.OrdinalIgnoreCase`.
- `ModifierKeysFlag` static field is now a compile-time constant, saving memory and providing speed-up.
- The JITTed codegen takes up less memory as we removed few things.
- Adds/completes code documentation for the whole class.

### `ConvertFrom` Benchmark for all possible modifiers

| Method   | source               | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size | Allocated |
|--------- |--------------------- |----------:|-----------:|------------:|-------:|----------:|----------:|
| Original | CTRL (...)+ ALT [24] | 101.55 ns |   1.790 ns |    1.587 ns | 0.0224 |       2,662 B |         376 B |
| PR__EDIT | CTRL (...)+ ALT [24] |  34.68 ns |   0.218 ns |    0.182 ns |      - |       1,891 B |             - |
| Original | Ctrl (...)+ Alt [28] | 147.59 ns |   1.616 ns |    1.512 ns | 0.0319 |       2,640 B |         536 B |
| PR__EDIT | Ctrl (...)+ Alt [28] |  35.19 ns |   0.483 ns |    0.451 ns |      - |       1,891 B |             - |

### `ConvertTo` benchmark for different scenarios
| Method    | value | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|---------- |------ |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original  | None  | 1.0984 ns |  0.0108 ns |   0.0090 ns |      - |         685 B |             - |
| PR__EDIT  | None  | 0.8665 ns |  0.0078 ns |   0.0066 ns |      - |         238 B |             - |
| Original  | Shift | 1.1785 ns |  0.0149 ns |   0.0132 ns |      - |         780 B |             - |
| PR__EDIT  | Shift | 0.8840 ns |  0.0209 ns |   0.0196 ns |      - |         255 B |             - |
| Original | 3     | 10.517 ns |  0.1322 ns |   0.1104 ns | 0.0043 |       1,134 B |          72 B |
| PR__EDIT | 3     |  9.546 ns |  0.1713 ns |   0.1518 ns | 0.0024 |         756 B |          40 B |
| Original  | 15    | 39.23 ns |   0.828 ns |    0.886 ns | 0.0176 |       1,137 B |         296 B |
| PR__EDIT  | 15    | 11.88 ns |   0.194 ns |    0.181 ns |  0.0043 |  753 B      |   72 B |

## Customer Impact

Increased performance, zero/reduced allocations, less static memory.

## Regression

No.

## Testing

Local build, CI, assert tests:
```csharp
//ConvertTo
string x1 = Original(ModifierKeys.None);
string x2 = PR__EDIT(ModifierKeys.None);
Assert.AreEqual(x1, x2);
x1 = Original(ModifierKeys.Shift);
x2 = PR__EDIT(ModifierKeys.Shift);
Assert.AreEqual(x1, x2);
x1 = Original(ModifierKeys.Shift | ModifierKeys.Alt);
x2 = PR__EDIT(ModifierKeys.Shift | ModifierKeys.Alt);
Assert.AreEqual(x1, x2);
x1 = Original(ModifierKeys.Control | ModifierKeys.Alt);
x2 = PR__EDIT(ModifierKeys.Control | ModifierKeys.Alt);
Assert.AreEqual(x1, x2);
x1 = Original(ModifierKeys.Windows | ModifierKeys.Alt);
x2 = PR__EDIT(ModifierKeys.Windows | ModifierKeys.Alt);
Assert.AreEqual(x1, x2);
x1 = Original(ModifierKeys.Windows | ModifierKeys.Alt | ModifierKeys.Control | ModifierKeys.Shift);
x2 = Original(ModifierKeys.Windows | ModifierKeys.Alt | ModifierKeys.Control | ModifierKeys.Shift);
Assert.AreEqual(x1, x2);

//ConvertFrom
ModifierKeys x1 = Original("CTRL + ");
ModifierKeys x2 = PR__EDIT("CTRL + ");
Assert.AreEqual(x1, x2);
x1 = Original("CTRL");
x2 = PR__EDIT("CTRL");
Assert.AreEqual(x1, x2);
x1 = Original(" ALT + SHIFT ");
x2 = PR__EDIT(" ALT + SHIFT ");
Assert.AreEqual(x1, x2);
x1 = Original("WIN + WINDOWS");
x2 = PR__EDIT("WIN + WINDOWS");
Assert.AreEqual(x1, x2);
x1 = Original("   ");
x2 = PR__EDIT("   ");
Assert.AreEqual(x1, x2);
x1 = Original("CTRL + WIN + SHIFT + ALT");
x2 = PR__EDIT("CTRL + WIN + SHIFT + ALT");
Assert.AreEqual(x1, x2);
x1 = Original("CTRL ++ WIN");
x2 = PR__EDIT("CTRL ++ WIN");
Assert.AreEqual(x1, x2);
x1 = Original("ALT+WIN");
x2 = PR__EDIT("ALT+WIN");
Assert.AreEqual(x1, x2);
```

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9683)