@harshit7962 2025-01-03
Also, as we already have `WindowsBase` project already set-up for unit tests, can we add some tests for this change?