@h3xds1nz 2025-01-03
@harshit7962 I didn't PR any as there are tests in #8215, same for KeyConverter. I've unfortunately found out after I wrote them myself, haha.

On my way with the conflicts.